Outsource router.
