+++ frontend/src/main.js
@@ -5,6 +5,7 @@ import Vue from "vue";
 import VueRouter from "vue-router";
 import vuetify from "@/plugins/vuetify"; // path to vuetify export
 import App from "./App.vue";
+import router from "./router";
 
 import { BaklavaVuePlugin } from "@baklavajs/plugin-renderer-vue";
 
@@ -15,10 +16,6 @@ import "highlight.js/styles/github.css";
 import VueExcelEditor from "vue-excel-editor";
 
 import store from "./vuex_store";
-import NodeEditor from "./components/NodeEditor.vue";
-import ResultsDashboard from "./components/ResultsDashboard.vue";
-import EstimatesDashboard from "./components/EstimatesDashboard.vue";
-import CodeDashboard from "./components/CodeDashboard.vue";
 
 Vue.use(VueRouter);
 Vue.use(BaklavaVuePlugin);
@@ -31,34 +28,6 @@ Vue.config.productionTip = false;
 Vue.config.devtools = false;
 
 Vue.prototype.log = console.log;
-
-const routes = [
-  {
-    path: "/",
-    redirect: "/modeling"
-  },
-  {
-    path: "/modeling",
-    component: NodeEditor
-  },
-  {
-    path: "/estimates",
-    component: EstimatesDashboard
-  },
-  {
-    path: "/results",
-    component: ResultsDashboard
-  },
-  {
-    path: "/code",
-    component: CodeDashboard
-  }
-];
-
-const router = new VueRouter({
-  routes, // short for `routes: routes`
-  mode: "history"
-});
 
 /* eslint-disable no-new */
 new Vue({
